requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-DotEnv==0.1.2
Flask-JWT==0.3.2
Flask-JWT-Extended==4.1.0
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.3
numpy==1.20.2
pandas==1.2.4
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
six==1.15.0
Werkzeug==1.0.1
